Updated bash and node config for osx and nvm users
The original bash info didn't work for me until I added the export part, not sure if it was just for me, but I do remember other mac users having this issue. The nvm part I feel needs to be here or somewhere in general as its pretty helpful and on of the first issues you'd experience with VS Code if you are using nvm.