Explain we have opened two VS Code instances
I was not a very dutiful reader and somewhat skimmed the initial commands which lead me to pointing one instance of VS Code at the entire cloned repository.

Later when I read,
> Go to the VS Code instance containing the server code (see above) and press [...]

I was lost because I did not remember reading anything about opening dedicated instances.  A quick text search for _instance_ did not find anything earlier in the document, so I had to go back and re-read.

This adds a brief explanatory sentence that might be caught by a reader or at least provides something to search for when you encounter the text "see above".