Use LF as the line-ending character consistently
Developers from different operating systems are facing inconsistent line
endings all the time, which causes a lot of pain.

This commit replaces all CRLFs with LFs to follow vscode's convention.

https://github.com/microsoft/vscode/blob/master/.vscode/settings.json

The new settings in .vscode/settings.json and .editorconfig also
enforces the use of LF as the line-ending character in future commits.