Fixing the broken hyperlink for "Java Extension Pack"
This PR fixes the URL for "Java Extension Pack" link which had a typo in the URL. This results in having a broken link on the following page: https://code.visualstudio.com/docs/java/java-debugging (search for "Java Extension Pack" on that page). It doesn't repro when the markdown file is viewed on GitHub.