Add fenced code block language identifier description
The Markdown Syntax page does not include a list of supported languages that can be used as code block language identifiers. At the moment the only way to get that list is by reading the source code, which is a bit of a pain.

These definitions were gleaned from `vscode/extensions/markdown/syntaxes/gulpfile.js` on the latest master branch so they are current.