Updated required .NET Framework version
When trying to run VS Code on a PC with Windows 7 (64 bit) that only had .NET Framework 4.5.2 installed (as per this guide), an error popped up when opening VS Code stating that the required .NET Framework version is 4.6. After updating .NET Framework on the PC, VS Code ran without issues. Hence the higher version requirement for the .NET Framework should be reflected in the documentation.