🐛 FIX: Grammar splitting the infinitive
There is no grammatical rule against splitting the infinitive, but some people strongly object to the construction, especially in formal writing. In many instances, unsplitting the infinitive can sound awkward or change the meaning of the sentence. If your reader is likely to object to the split infinitive, rewrite the sentence to avoid its use. (I flag this for that reason.) It is unlikely that long phrases and interrupters can smoothly split the infinitive.