Update docs on cmder
The comment in the vscode.bat snippet causes the batch file to not work. This can cause confusion for beginners who may think it is okay to leave it in there, and may not understand what they are doing wrong. My update takes the comment out of the batch file.