Update Go extension screenshot
Since the Go extension changed publisher from lukehoban to Microsoft, it
could be confusing for users seeing one publisher in the docs and a
different one in the Extensions tab. Some might also think that the two
are different extensions.

This commit updates the screenshot go-extension.png with an up-to-date
one.