Fix description of the debugging process
In the original text it refers to the debug "continue" command as the "start" command but those are different commands.  Once you start debugging, F5 is assigned to "continue" so the description has been fixed to change "start" to "continue" in that section and to remove the reference to pressing the green arrow (which generates an error message to the user saying that a debug session is already underway).