Update .NET Core download link
Updating to the main .NET download page. The current link lands folks in a tutorial to build a hello world app with .NET Core CLI. We've heard from customers (via the Feedback button on the .NET site) that this is confusing.

The target page will automatically detect OS and show the correct downloads.