Click is not accessible language
Came up in another doc I am reviewing (that references this one) that this doc uses "click." Click is not an accessible term (because it implies that you are using a mouse and can click it). Select is the preferred MSFT voice term because it's neutral about how a person goes about doing the action but implies the same end action. See Microsoft Voice/Style guidelines for more details.