Update python/debugging.md debugOptions
The changes to `debugOptions` in https://github.com/Microsoft/vscode-python/issues/1326 (which are reflected in the behaviour of my install of VS Code so I assume have been incorporated into the main debugger) do not appear to be reflected in the documentation.  This was causing me some problems.

So have updated the file to reflect them.