Recommend to always use strict host key checking
I've searched for a long time before finding the cause of my timout at "Waiting for ssh tunnel to be ready". So I thought I'd share here. Not sure if this is a bug or something the users should be aware of. It would be nice if vscode intercept that use case where the  target host key has changed (ie.: "Port forwarding is disabled to avoid man-in-the-middle attacks." warning). Let me know if a enhancement request would be a better way to go here.
Also, as an addition to what I put here, it would be nice to provide an external guide on how to properly add trusted host key. I didn't include one as I don't know if :
1- you already have one somewhere that I'm not aware of
2- what is your policy for external link in the documentation

Thanks for considering this!