Correct cmd command for matplotlib pkg installment
Using the integrated terminal within vscode, the cmd ` python -m pip install matplotlib` generated the following error

```
python : The term 'python' is not recognized as the name of a cmdlet, function, script file, 
or operable program. Check the spelling of the name, or if a path was included, verify that 
the path is correct and try again.
At line:1 char:1
+ python -m pip install matplotlib
+ ~~~~~~
    + CategoryInfo          : ObjectNotFound: (python:String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException
```

but the command `py -m pip install matplotlib` work as expected.