Limit the validity of the signing key to packages.microsoft.com
Before this change the key has been trusted for all possible repositories.
Moving it to /usr/share/keyrings/, the conventional location, prevents that.

The 'signed-by' links location with its particular, authoritative key(s).

To prevent any name clashings, it's a good practice to not use an
organizations name here (except the key or keyring is really meant to cover
the entire org, like GNU's) and go with with a descriptive one,
or one obvious to where it will be used.