Fix incoherence in Virtual Documents guide
The virtual documents guide talks about event emitters but does not actually use them in the code example. 
The text has been updated to reflect what the code does, and the mention of the event emitter feature is now accompanied by a link to an example making use of it.