not clear on when initializeCommand runs
When reading this first, it seems as if `initializeCommand` would run only before container image being built. But it clearly runs both when first build of the image AND when re-attaching to an existing stopped container.  Maybe to others its clear, but to me it wasn't.

I'd actually would like a step that is "just before an image build" :)