RAM requirements much bigger?
I googled the RAM requirements a bit and there are posts on Reddit and elsewhere mentioning how big this 1GB requirement is for Code.. Yet for me it is 3.8GB, not 1. 

It likely depends on the programs being run: I'm playing with an Typescript, React, Prettier and ESLint app. As many I made the switch to Code because of TS, and the other tools there are not expendable. 

Shall one deduce that the requirement is in reality 4GB?