clarify how to disable pytest-cov when debugging (#3770)
* clarify how to disable pytest-cov when debugging

The previous form, "To prevent this behavior, include --no-cov in pytestArgs when debugging tests.", left ambiguous _how_ to modify the pytestArgs setting "when debugging tests".
Let's take advantage of the earlier-explained launch.json config for debugging tests. We can't directly change Visual Studio Code settings there, but we can set the env var PYTEST_ADDOPTS, which will have much the same effect. With this snippet, test-running just "does the right thing": coverage on Run Tests, and debugging on Debug Tests.

As to _how_ to provide the snippet, I defaulted to just adding a subordinate clause, for minimal disruption to the existing text. I'm not entirely happy with how verbose it makes the paragraph.

* be clearer about the debug-tests link

Co-authored-by: Eric Snow <ericsnowcurrently@gmail.com>

Co-authored-by: Eric Snow <ericsnowcurrently@gmail.com>