added troubleshooting for no input terminal
There is a huge issue with the input entry with the newer versions of Mac. I spent at least 2 hours trying to understand why this is happening. 

Until the issue is fixed, I think, this troubleshooting must be included in the main documentation as this is where new developers, especially students would land, when trying to run a simple hello world on their new Macs.