On branch edburns-msft-workspaces Suggested improvements modified:   docs/editor/workspaces.md
- Change wording used for "VS Code window (instance)" to "running
  instance of VS Code".

- Mention what we mean about the term "project".

- Change "and" to ",".

## Multi-root workspaces

This section is focused on opening a *.code-workspace* file.  I've
worked with multi-root workspaces a long time and never once have I felt
the need to open one of those files.  Rather, I use the UI to add/remove folders
from my multi-root workspaces.  Introducing the editing of this file in
this document is confusing and, in my opinion, unnecessary.  I really
would rather delete that file related content from this document
alltogether, but that is too big a change for me to suggest in a PR from
some rando.  Instead, I'll add text to put the *.code-workspace* content
in its proper perspective.

- Add note

+Multi-root workspaces are very useful to implement the loose definition of "project" stated above.

- The word workspace is in all caps in the UI.

- Add indefinite article.