Update to ReadMe
I helped a friend on a Windows PC running through this installation, and she did not select win_32 for threads, but kept the POSIX option. Could be nice to clarify this. 
I am on a Mac so can't confirm that "win_32" is the exact name, but am pretty sure it was/is.