Update index page to fix an issue in section list
The sections listed in the section "How to build extensions" had a formatting issue, i.e it had an extra asterisk in addition to the two that were required to boldify the section title - for example, "* ***Get Started** teaches fundamental concepts". This PR fixes the issue by removing that extra asterisk and make it as "* **Get Started** teaches fundamental concepts".