Move tutorials to top of doc
Can we move the tutorials section to the top of this doc so that people can more easily find the content for their platform?

BTW, we're going to start moving away from recommending MinGW for customers who want to target Linux from Windows and instead recommend WSL for those workflows. I'm going to propose changes to the GCC on Windows doc, and then we'll probably want to pick a different example for this landing page. I can help with that.