Correct quickstart-python.md prompt order
Updated the order of the prompts after selecting `Python: Django` as the app type when using `Docker: Add Docker Files to Workspace...`. For the Docker extension (v1.13.0), the correct order should have the prompt to select whether or not to include Docker Compose files as the last prompt, not the first.