settings: clarify editor can't write defaults
The interactive settings editor will not write a default value to the workplace settings file. While there are good reasons for this, it is somewhat confusing, as many have noted on https://github.com/microsoft/vscode/issues/58038. This PR documents the current behavior and suggests the workaround of manually editing the workspace settings file.