docs: Update the Tree file filtering doc section in getStarted userInterface
The find control tree filtering was implemented and is the current. And the old way of tree file filtering through focus > type is deprecated. 

This PR updates the Tree file filtering so it reflects the new changes. As it can be confusing if you read the documentation. And you get something that doesn't work.

- updated the section with the release note. As it's complete and clear.
- Added a note about deprecation and what version. With a link to the release note.
- All links and images are mapped internally with relative paths.