I think that port 3000 should be forwarded.
When I tried this tutorial, I could not open 'http://localhost:3000' from a host web browser.
I noticed that the port 3000 of the container was not published by executing 'docker ps' on the host.
Therefore, I think that port 3000 should be forwarded at 'devcontainer.json' (and '9000' at "portsAttributes" section may be wrong...)
And the port 3000 should be opend at the image of container, shouldn't it?