Document the fourth param to vscode.diff
This being missing from the docs cost me an hour or two! I only found it when I searched for other code using the API and found that about half of callers passed this fourth option.

While it's called columnOrOptions, I've only been able to get it to work with a vscode.TextDocumentShowOptions.